Fix a whole bunch of warnings and one actual bug (assignment inside conditional in textoken.c) that are reported by "cl /W4" (thanks Javier Mugica)
git-svn-id: https://serveur-svn.lri.fr/svn/modhel/luatex/trunk@3354 0b2b3880-5936-4365-a048-eb17d2e5a6bf

+++ source/texk/web2c/luatexdir/tex/arithmetic.c
@@ -217,8 +217,9 @@ scaled xn_over_d(scaled x, int n, int d)
         tex_remainder = v % d;
         return u;
     } else {
-        tex_remainder = -(v % d);
-        return -u;
+        /* casts are for ms cl */
+        tex_remainder = -(int)(v % d);
+        return -(int)(u);
     }
 }
